requirements.txt
Django==2.2.28
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.6
sqlparse==0.3.1 
apachelogs==0.6.0
django-crontab==0.7.1
python-dotenv==0.19.2
python-magic==0.4.27
drf-spectacular==0.26.2